requirements.txt
asgiref==3.7.2
Brotli==1.0.9
certifi==2023.5.7
charset-normalizer==3.1.0
crispy-bootstrap4==2022.1
dj-database-url==2.0.0
Django==4.2.2
django-countries==7.5.1
django-crispy-forms==2.0
django-embed-video==1.4.8
gunicorn==20.1.0
idna==3.4
Pillow==9.5.0
psycopg2==2.9.6
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
whitenoise==6.5.0
